mobile: Rename common/common directory
