+++ mobile/library/common/BUILD
@@ -25,11 +25,11 @@ envoy_cc_library(
     deps = [
         ":engine_common_lib",
         "//library/common/bridge:utility_lib",
-        "//library/common/common:logger_delegate_lib",
         "//library/common/data:utility_lib",
         "//library/common/event:provisional_dispatcher_lib",
         "//library/common/http:client_lib",
         "//library/common/http:header_utility_lib",
+        "//library/common/logger:logger_delegate_lib",
         "//library/common/network:connectivity_manager_lib",
         "//library/common/stats:utility_lib",
         "//library/common/types:c_types_lib",
