+++ mobile/library/common/http/BUILD
@@ -13,7 +13,6 @@ envoy_cc_library(
     deps = [
         "//library/common/bridge:utility_lib",
         "//library/common/buffer:bridge_fragment_lib",
-        "//library/common/common:system_helper_lib",
         "//library/common/data:utility_lib",
         "//library/common/event:provisional_dispatcher_lib",
         "//library/common/extensions/filters/http/local_error:local_error_filter_lib",
@@ -22,6 +21,7 @@ envoy_cc_library(
         "//library/common/network:connectivity_manager_lib",
         "//library/common/network:synthetic_address_lib",
         "//library/common/stream_info:extra_stream_info_lib",
+        "//library/common/system:system_helper_lib",
         "//library/common/types:c_types_lib",
         "@envoy//envoy/buffer:buffer_interface",
         "@envoy//envoy/common:scope_tracker_interface",
