+++ mobile/library/common/internal_engine.h
@@ -7,9 +7,9 @@
 
 #include "absl/base/call_once.h"
 #include "extension_registry.h"
-#include "library/common/common/logger_delegate.h"
 #include "library/common/engine_common.h"
 #include "library/common/http/client.h"
+#include "library/common/logger/logger_delegate.h"
 #include "library/common/network/connectivity_manager.h"
 #include "library/common/types/c_types.h"
