+++ mobile/library/common/logger/logger_delegate.h
