+++ mobile/library/common/system/BUILD
@@ -61,18 +61,3 @@ envoy_cc_library(
     repository = "@envoy",
     deps = [],
 )
-
-envoy_cc_library(
-    name = "logger_delegate_lib",
-    srcs = ["logger_delegate.cc"],
-    hdrs = ["logger_delegate.h"],
-    repository = "@envoy",
-    deps = [
-        "//library/common/api:external_api_lib",
-        "//library/common/bridge:utility_lib",
-        "//library/common/data:utility_lib",
-        "//library/common/types:c_types_lib",
-        "@envoy//source/common/common:macros",
-        "@envoy//source/common/common:minimal_logger_lib",
-    ],
-)
