+++ mobile/library/common/system/default_system_helper.cc
@@ -1,4 +1,4 @@
-#include "library/common/common/default_system_helper.h"
+#include "library/common/system/default_system_helper.h"
 
 namespace Envoy {
