+++ mobile/test/common/logger/BUILD
@@ -10,8 +10,8 @@ envoy_cc_test(
     repository = "@envoy",
     deps = [
         "//library/common/api:external_api_lib",
-        "//library/common/common:logger_delegate_lib",
         "//library/common/data:utility_lib",
+        "//library/common/logger:logger_delegate_lib",
         "//library/common/types:c_types_lib",
     ],
 )
