+++ mobile/test/common/mocks/common/mocks.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "gmock/gmock.h"
-#include "library/common/common/system_helper.h"
+#include "library/common/system/system_helper.h"
 
 namespace Envoy {
 namespace test {
